Title: 정수 사각형 최대 합, Time: 9ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 240226/정수 사각형 최대 합/maximum-sum-path-in-square.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int map[100][100];
-int dp[100][100];
+long long dp[100][100];
 int n;
 void solve(){
     dp[1][1] = map[1][1];
